+++ src/main/java/com/deviget/minesweeper/infra/service/CellOperation.java
@@ -1,0 +1,7 @@
+package com.deviget.minesweeper.infra.service;
+
+public enum CellOperation {
+  UNCOVERING,
+  FLAG_PLACEMENT,
+  QUESTION_MARK_PLACEMENT
+}