+++ src/main/java/com/deviget/minesweeper/infra/service/GameService.java
@@ -2,14 +2,21 @@ package com.deviget.minesweeper.infra.service;
 
 import com.deviget.minesweeper.domain.model.game.Game;
 import com.deviget.minesweeper.domain.model.game.GameId;
+import com.deviget.minesweeper.domain.model.game.GameStatus;
 import com.deviget.minesweeper.domain.model.game.Preferences;
+import com.deviget.minesweeper.domain.model.game.cell.CellCoordinate;
 import com.deviget.minesweeper.domain.model.user.UserId;
 import com.deviget.minesweeper.infra.repository.GameRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class GameService {
+
+  @Autowired
+  private CellOperations cellOperations;
 
   @Autowired
   private GameRepository gameRepository;
@@ -19,18 +26,19 @@ public class GameService {
   }
 
   public Game resume(final GameId gameId, final UserId userId) {
-    final Game existingGame =
-            gameRepository.findByIdAndUserId(gameId, userId);
-
+    final Game existingGame = gameRepository.findByIdAndUserId(gameId, userId);
     existingGame.resume();
-
-    gameRepository.save(existingGame);
-
-    return existingGame;
+    return gameRepository.save(existingGame);
   }
 
-  public Game find(final GameId gameId, final UserId userId) {
-    return gameRepository.findByIdAndUserId(gameId, userId);
+  public Game performCellOperation(final GameId gameId,
+                                   final UserId userId,
+                                   final CellOperation operation,
+                                   final CellCoordinate cellCoordinate) {
+    return cellOperations.of(operation).perform(gameId, userId, cellCoordinate);
   }
 
+  public List<Game> findUnfinished(final UserId userId) {
+    return gameRepository.findByUserIdAndStatusIn(userId, GameStatus.unfinished());
+  }
 }